Added Custom Lists
Created a functionality to add any number of your own custom lists.
The database is hosted locally, so is the server as of this commit.

+++ todolist-v1/app.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const bodyParser = require('body-parser');
 const date = require(__dirname + '/date.js');
+const _ = require('lodash');
 const mongoose = require('mongoose');
 
 const app = express();
@@ -8,7 +9,9 @@ app.use(bodyParser.urlencoded({ extended: true }));
 app.set("view engine", "ejs");
 app.use(express.static('public'));
 
-mongoose.connect("mongodb://127.0.0.1:27017/todolistDB");
+const port = process.env.PORT || 3000;
+
+mongoose.connect("mongodb+srv://admin-DTS:DTS123@cluster0.enq3fzf.mongodb.net/todolistDB");
 
 const itemSchema = new mongoose.Schema({
     name: String
@@ -17,22 +20,25 @@ const itemSchema = new mongoose.Schema({
 const Item = new mongoose.model("item", itemSchema);
 
 const todoItem1 = new Item({
-    name: "Do Homework"
+    name: "Type Below to add Items"
 })
 
 const todoItem2 = new Item({
-    name: "DSA"
+    name: "<-- Hit this to delete"
 })
 
 var defaultItems = [todoItem1, todoItem2];
 const options = {ordered: true};
 
+const listSchema = new mongoose.Schema({
+    name: String,
+    items: [itemSchema]
+})
 
-var workItem = '';
-var workItems = [];
+const List = new mongoose.model("list", listSchema);
 
-app.listen(3000, function (req, res) {
-    console.log("Server running on port 3000");
+app.listen(port, function (req, res) {
+    console.log(`Server running on port ${port}`);
 })
 
 app.get('/', function (req, res) {
@@ -44,7 +50,7 @@ app.get('/', function (req, res) {
             Item.insertMany(defaultItems, options);
             res.redirect('/');
         }
-        else{
+        else{            
             res.render('list', { 
                 listType: day, 
                 list_items: items 
@@ -58,39 +64,78 @@ app.get('/', function (req, res) {
 
 app.post('/', function(req, res){
     const itemName = req.body.next_item;
+    const listName = req.body.list;
 
     const todoItem = new Item({
         name: itemName
     })
-    todoItem.save();
-    console.log("Successfully inserted");
 
-    if(req.body.list == 'Work List'){
-        workItems.push(workItem);
-        res.redirect('/work');
+    if(listName == date.getDate()){
+        todoItem.save();
+        res.redirect('/');
+    }
+    else{
+        List.findOne({name: listName})
+        .then((list) => {
+            list.items.push(todoItem);
+            list.save();
+            res.redirect("/list/" + listName);
+        })
+        .catch(err => console.log(err))
     }
     
-    res.redirect('/');
 })
 
-app.post("/delete", function(req, res){
-    const del = req.body.deletedItem;
+app.post("/delete/:listName", function(req, res){
+    const del = req.body.deletedItem;        
+    const listTitle = req.params.listName;
     
-    Item.deleteOne({_id: del})
-    .then(() => {
-        console.log("Successfully deleted")
+    if(listTitle == date.getDate()){
+        Item.deleteOne({_id: del})
+        .then(() => {
+            console.log("Successfully deleted")
+        })
+        .catch(err => console.log(err));
+
+        res.redirect("/");
+    }
+    else{
+        List.updateOne({name : listTitle},{$pull: {items: {_id: del}}})
+        .then(() => res.redirect("/list/" + listTitle))
+        .catch(err => console.log(err));        
+    }
+})
+
+app.get('/list/:listName', function(req, res){
+    const listTitle = _.capitalize(req.params.listName);
+    
+
+    List.findOne({name: listTitle})
+    .then((list) => {
+        if(!list){
+            //Create a new list            
+            const list = new List({
+                name: listTitle,
+                items: []
+            })
+            list.save();
+
+            res.redirect("/list/" + listTitle);
+        }
+        else{
+            //append into existing list            
+            res.render('list', {
+                listType : listTitle, 
+                list_items: list.items
+            })
+        }
     })
-    .catch(err => console.log(err));
+    .catch(err => {
+        console.log(err);
+    })
 
-    res.redirect("/");
-})
 
-app.get('/work', function(req, res){
-    res.render('list', {
-        listType: 'Work List',
-        list_items: workItems
-    });
-})
+});
 
 app.get('/about',function(req, res){
     res.render('about');
